Refactor Header.jsx
