+++ src/helpers/get-header-paths.js
