+++ src/page-renderers/Components/Header.jsx
@@ -1,5 +1,5 @@
 const React = require('react')
-const getHeaderLinks = require('../../helpers/get-header-links')
+const getHeaderPaths = require('../../helpers/get-header-paths')
 
 const Styles = {
   navbar: {
@@ -19,11 +19,16 @@ const Styles = {
 
 const Header = ({ relPath }) => {
   let pathParts
+  let headerPaths
   let headerLinks
 
   if (relPath) {
     pathParts = relPath.split('/')
-    headerLinks = getHeaderLinks(relPath)
+    headerPaths = getHeaderPaths(relPath)
+
+    headerLinks = headerPaths.map((dir, index) => {
+      return <a style={{ textAlign: 'left', paddingRight: '3px', color: '#4f8df0', fontSize: '20' }} href={'/' + dir} key={index}><span>{'/' + pathParts[index]}</span></a>
+    })
   }
 
   return (
@@ -31,11 +36,7 @@ const Header = ({ relPath }) => {
       <div className='row col-md-12' style={{ width: '100%', paddingTop: '10px', paddingRight: '0px', paddingLeft: '25px' }}>
         <div className='col-md-4' style={Styles.headerLinks}>
           {
-            headerLinks
-              ? (() => headerLinks.map((dir, index) => {
-                return <a style={{ textAlign: 'left', paddingRight: '3px', color: '#4f8df0', fontSize: '20' }} href={'/' + dir} key={index}><span>{'/' + pathParts[index]}</span></a>
-              }))()
-              : <h3 style={{ color: '#fff', marginTop: '5px' }}>Morty-Docs</h3>
+            headerLinks ? headerLinks : <h3 style={{ color: '#fff', marginTop: '5px' }}>Morty-Docs</h3>
           }
         </div>
       </div>
